Skip test as we are not going to customize the RemoteCensusApi integration
Instead of customizing the Census integration using RemoteCensusApi
class we used CensusApi. As we do not have plans to use this one we can
safely ignore it.

The error comes because we customized the data coming from the Census
webservice, and we are saving it in the user. All these new user data is
not return by the default remote_census_api implementation. If at some
point we want to make it work we will need to update the RemoteCensusApi
so it works similar to the custom CensusApi we created.