Fix deployment node prefix when using multiple servers
When Capistrano runs on multiple hosts, it prepends this prefix for
every host, so you end up with:

```
EXECJS_RUNTIME='' $HOME/.fnm/fnm exec EXECJS_RUNTIME='' $HOME/.fnm/fnm
exec npm install ...
```

This is not a valid command, and SSHKit tries to execute
`EXECJS_RUNTIME=`, which fails.

By using a string, not a lambda, for the prefix, and ensuring it is not
prepended multiple times.